Orderflow/Backend/requirements.txt
# Django Core
Django>=5.0,<7.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0
PyJWT>=2.8,<3.0

# Database
psycopg2-binary>=2.9,<3.0

# Environment & Security
python-dotenv>=1.0,<2.0
python-decouple>=3.8,<4.0

# Utilities
Pillow>=10.0,<13.0
django-filter>=23.0,<24.0
drf-spectacular>=0.27,<1.0

# QR Code Generation
qrcode>=7.4,<9.0

# Payment Gateway
razorpay>=1.4,<3.0

# Development
django-debug-toolbar>=4.2,<7.0

# Testing
pytest>=7.4,<9.0
pytest-django>=4.7,<5.0
factory-boy>=3.3,<4.0

# Production
gunicorn>=21.0,<24.0
whitenoise>=6.6,<7.0


# Async & background jobs
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.6,<3.0